Refactor edit functionality
